[refactor] CrimeView의 type 개별 선언 삭제
- CrimeView의 type이 개별로 선언되어있는 것을 삭제
   src/types의 type으로 통일
