+++ src/components/IssueDetail/CrimeView/CrimeHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box, createStyles, makeStyles, Typography, Theme } from '@material-ui/core';
 
-import { ICrime } from './types';
+import { ICrime } from '../../../types';
 import BackNextButtons from '../../NewProject/BackNextButtons';
 
 interface IProps {
